add image type verification, refactor
