+++ batch-processing/helpers.go
@@ -4,11 +4,11 @@ import (
 	"encoding/csv"
 	"fmt"
 	"io"
+	"io/ioutil"
 	"log"
 	"net/http"
 	"os"
 	"strings"
-	"sync"
 	"time"
 
 	"errors"
@@ -17,6 +17,10 @@ import (
 const (
 	InvalidCSVFormat   = "CSV file must have a header row with 'url' as the first column"
 	CouldNotFetchImage = "Received status %d when trying to download image"
+)
+
+var (
+	SupportedImageTypes = []string{"imag1e/jpeg", "image/png"}
 )
 
 func ReadCSV(inputFilepath *string) *csv.Reader {
@@ -51,6 +55,11 @@ func downloadFile(URL, filePath string) error {
 	if response.StatusCode != http.StatusOK {
 		return errors.New(fmt.Sprintf(CouldNotFetchImage, response.StatusCode))
 	}
+
+	err = verifyImageType(&response.Body)
+	if err != nil {
+		return err
+	}
 	//Create a empty file
 	file, err := os.Create(filePath)
 	if err != nil {
@@ -67,31 +76,6 @@ func downloadFile(URL, filePath string) error {
 	return nil
 }
 
-func ConvertFile(url string, c Converter, r chan *Output, wg *sync.WaitGroup) {
-	defer wg.Done()
-	fileExt := "jpg"
-	fileName := extractFilename(url)
-	inputPath := fmt.Sprintf("/outputs/%s.%s", fileName, fileExt)
-	outputPath := fmt.Sprintf("/outputs/%s-%s.%s", fileName, "converted", fileExt)
-
-	fmt.Println("Downloading image from URL: ", url)
-	err := downloadFile(url, inputPath)
-
-	if err != nil {
-		r <- &Output{url: url, err: err}
-	}
-
-	// Do the conversion!
-	err = c.Grayscale(inputPath, outputPath)
-
-	if err != nil {
-		r <- &Output{url: url, input: inputPath, output: outputPath, err: err}
-		return
-	}
-
-	r <- &Output{url: url, input: inputPath, output: outputPath}
-}
-
 func extractFilename(url string) string {
 	urlArr := strings.Split(url, "/")
 
@@ -104,37 +88,27 @@ func extractFilename(url string) string {
 	return fileName
 }
 
-func ResultTOCSV(rows <-chan *Output, outputFilepath, failedOutputFilepath string) {
+func verifyImageType(reader *io.ReadCloser) error {
+	bytes, err := ioutil.ReadAll(*reader)
 
-	f, err := os.Create(outputFilepath)
 	if err != nil {
-		log.Fatalf("error creating output file: %v", err)
+		return errors.New(fmt.Sprintf("Error reading downloaded image: %v", err))
 	}
-	defer f.Close()
 
-	w := csv.NewWriter(f)
-	w.Write([]string{"url", "input", "output", "s3url"})
-	var failedW *csv.Writer
-	defer w.Flush()
+	mimeType := http.DetectContentType(bytes)
 
-	if failedOutputFilepath != "" {
-		failedF, err := os.Create(failedOutputFilepath)
-		if err != nil {
-			log.Fatalf("error creating failed output file: %v", err)
+	if !contains(SupportedImageTypes, mimeType) {
+		return errors.New(fmt.Sprintf("Unsupported image type: %s, only support: %s", mimeType, SupportedImageTypes))
+	}
+
+	return nil
+}
+
+func contains(s []string, e string) bool {
+	for _, a := range s {
+		if a == e {
+			return true
 		}
-		defer failedF.Close()
-
-		failedW = csv.NewWriter(failedF)
-		failedW.Write([]string{"url", "input", "output", "s3url", "error"})
-		defer failedW.Flush()
 	}
-
-	for row := range rows {
-		if row.err != nil && failedOutputFilepath != "" {
-			failedW.Write([]string{row.url, row.input, row.output, row.s3url, row.err.Error()})
-			continue
-		}
-
-		w.Write([]string{row.url, row.input, row.output, row.s3url})
-	}
+	return false
 }
